requirements.txt
# Django Stuff
Django==4.2.10
djangorestframework==3.14.0

#Postgrest
psycopg2==2.9.9


#PDF
boto3==1.34.54
xhtml2pdf==0.2.15


